minor revisions

+++ cw04/zad3/main.c
@@ -6,45 +6,44 @@
 #include <unistd.h>
 #include <sys/wait.h>
 
+int CONF = 0;
 int L = 0;
 int TYPE = 0;
-int S_COUNT = 0;
-int R_COUNT = 0;
+int P_COUNT = 0;
+int C_COUNT = 0;
 
-void p_sigusr1_handler(int signum, siginfo_t* sig_info, void* sig_context){
-    S_COUNT++;
+void p_sig_handler(int signum, siginfo_t* sig_info, void* sig_context){
+    P_COUNT++;
+    CONF = 1;
 }
 
-void c_sigusr1_handler(int signum, siginfo_t* sig_info, void* sig_context){
-    R_COUNT++;
-    kill(getppid(), SIGUSR1);
+void c_sig_handler(int signum, siginfo_t* sig_info, void* sig_context){
+    C_COUNT++;
+    if (TYPE < 3) kill(getppid(), SIGUSR1);
+    else kill(getppid(), SIGRTMIN);
 }
 
-void c_sigusr2_handler(int signum, siginfo_t* sig_info, void* sig_context){
-    printf("SIGUSR1 signals received by child: %d\n", R_COUNT);
-    exit(0);
-}
-
-void p_rt1_handler(int signum, siginfo_t* sig_info, void* sig_context){
-    S_COUNT++;
-}
-
-void c_rt1_handler(int signum, siginfo_t* sig_info, void* sig_context){
-    R_COUNT++;
-    kill(getppid(), SIGRTMIN);
-}
-
-void c_rt2_handler(int signum, siginfo_t* sig_info, void* sig_context){
-    printf("SIGRTMAX signals received by child: %d\n", R_COUNT);
+void c_exitsig_handler(int signum, siginfo_t* sig_info, void* sig_context){
+    char* buffer = calloc(100, 1);
+    if (TYPE < 3){
+        sprintf(buffer, "SIGUSR1 signals received by child: %d\n", C_COUNT);
+        write(1, buffer, 100);
+    }
+    else{
+        sprintf(buffer, "SIGRTMAX signals received by child: %d\n", C_COUNT);
+        write(1, buffer, 100);
+    }
+    free(buffer);
     exit(0);
 }
 
 int parse(int argc, char* argv[]){
+    if (argc < 3) return -1;
     char * pEnd;
     L =  strtol (argv[1], &pEnd, 10);
     TYPE =  strtol (argv[2], &pEnd, 10);
-    if (L == 0 || TYPE == 0) return -1;
-    else if (TYPE != 1 && TYPE != 2 && TYPE != 3) return -1;
+    if (L == 0) return -1;
+    else if (TYPE <= 0 || TYPE > 3) return -1;
     else return 0;
 }
 
@@ -61,41 +60,29 @@ void print_error(int error_code){
 }
 
 void set_child_handlers(){
-    if (TYPE == 1 || TYPE == 2){
-        struct sigaction act;
-        sigfillset(&act.sa_mask);
-        act.sa_flags = SA_SIGINFO | SA_RESTART;
-        act.sa_sigaction = &c_sigusr1_handler;
+    struct sigaction act;
+    sigfillset(&act.sa_mask);
+    act.sa_flags = SA_SIGINFO | SA_RESTART | SA_NODEFER;
+    act.sa_sigaction = &c_sig_handler;
+    if (TYPE < 3){
         sigaction(SIGUSR1, &act, NULL);
-        act.sa_sigaction = &c_sigusr2_handler;
+        act.sa_sigaction = &c_exitsig_handler;
         sigaction(SIGUSR2, &act, NULL);
     }
     else{
-        struct sigaction act;
-        sigfillset(&act.sa_mask);
-        act.sa_flags = SA_SIGINFO | SA_RESTART;
-        act.sa_sigaction = &c_rt1_handler;
         sigaction(SIGRTMIN, &act, NULL);
-        act.sa_sigaction = &c_rt2_handler;
+        act.sa_sigaction = &c_exitsig_handler;
         sigaction(SIGRTMAX, &act, NULL); 
     }
 }
 
 void set_parent_handlers(){
-    if (TYPE == 1 || TYPE == 2){
-        struct sigaction act;
-        sigfillset(&act.sa_mask);
-        act.sa_flags = SA_SIGINFO | SA_RESTART;
-        act.sa_sigaction = &p_sigusr1_handler;
-        sigaction(SIGUSR1, &act, NULL);
-    }
-    else{
-        struct sigaction act;
-        sigfillset(&act.sa_mask);
-        act.sa_flags = SA_SIGINFO | SA_RESTART;
-        act.sa_sigaction = &p_rt1_handler;
-        sigaction(SIGRTMIN, &act, NULL);
-    }
+    struct sigaction act;
+    sigfillset(&act.sa_mask);
+    act.sa_flags = SA_SIGINFO | SA_RESTART | SA_NODEFER;
+    act.sa_sigaction = &p_sig_handler;
+    if (TYPE < 3) sigaction(SIGUSR1, &act, NULL);
+    else sigaction(SIGRTMIN, &act, NULL);
 }
 
 void send_signals1(){
@@ -104,26 +91,26 @@ void send_signals1(){
     if ((pid = fork()) > 0){
         set_parent_handlers();
         for (int i = 0; i < L; i++){
-            if (TYPE == 1) kill(pid, SIGUSR1);
-            else if (TYPE == 2){
+            if (TYPE == 3) kill(pid, SIGRTMIN);
+            else {
+                if (TYPE == 2) CONF = 0;
                 kill(pid, SIGUSR1);
-                pause();
-            }
-            else{
-                kill(pid, SIGRTMIN);
+                if (TYPE == 2){
+                    while (CONF == 0){}
+                }
             }
         }
-        if (TYPE == 1 || TYPE == 2){
+        if (TYPE < 3){
             printf("SIGUSR1 signals sent to child: %d\n", L);
             kill(pid, SIGUSR2);
             while (wait(NULL) > 0) {}
-            printf("SIGUSR1 signal responses received by parent: %d\n", S_COUNT);
+            printf("SIGUSR1 signal responses received by parent: %d\n", P_COUNT);
         }
         else{
             printf("SIGRTMIN signals sent to child: %d\n", L);
             kill(pid, SIGRTMAX);
             while (wait(NULL) > 0) {}
-            printf("SIGRTMIN signal responses received by parent: %d\n", S_COUNT);
+            printf("SIGRTMIN signal responses received by parent: %d\n", P_COUNT);
         }
     }
     else if (pid == 0){
